refactor: split starter.test.ts into parallelizable test files
- Extract defineTest utility to starter.ts
- Split tests into logical groups:
  - starter-basic.test.ts: Basic starter tests (dev/build for default and cloudflare)
  - starter-no-ssr.test.ts: No-SSR specific tests with build verification
  - starter-env.test.ts: Environment-specific tests (production/development modes)
  - starter-react-compiler.test.ts: React compiler integration tests
  - starter-base.test.ts: Base path configuration tests
  - starter-module-runner.test.ts: Module runner HMR configuration tests
  - starter-render-built-url.test.ts: renderBuiltUrl configuration tests

Benefits:
- Tests can now run in parallel for faster CI execution
- Better organization by functionality/configuration
- Easier maintenance with focused test files
- Maintains backward compatibility with original starter.test.ts

To run all starter tests:
pnpm test-e2e starter-basic starter-no-ssr starter-env starter-react-compiler starter-base starter-module-runner starter-render-built-url
